+++ app/src/test/java/nl/tudelft/b_b_w/blockchain/AcquaintanceTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.b_b_w.blockchaincomponents;
+package nl.tudelft.b_b_w.blockchain;
 
 import net.i2p.crypto.eddsa.EdDSAPrivateKey;
 import net.i2p.crypto.eddsa.EdDSAPublicKey;
@@ -7,14 +7,16 @@ import net.i2p.crypto.eddsa.Utils;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.util.ArrayList;
+
 import nl.tudelft.b_b_w.controller.ED25519;
 
 import static junit.framework.Assert.assertEquals;
 
 /**
- * This class tests the User object.
+ * Testing class for Acquaintance.java
  */
-public class UserTest {
+public class AcquaintanceTest {
 
     /**
      * The user attributes
@@ -22,7 +24,9 @@ public class UserTest {
     private String name;
     private String iban;
     private EdDSAPublicKey publicKey;
-    private User user;
+    private ArrayList<Chain> multiChain;
+
+    private Acquaintance acquaintance;
 
     /**
      * Setting up the test object
@@ -33,34 +37,15 @@ public class UserTest {
         iban = "NL642335674446";
         EdDSAPrivateKey edDSAPrivateKey = ED25519.generatePrivateKey(Utils.hexToBytes("0000000000000000000000000000000000000000000000000000000000000000"));
         publicKey = ED25519.getPublicKey(edDSAPrivateKey);
-        user = new User(name, iban, publicKey);
+
+        multiChain = new ArrayList<Chain>();
+        acquaintance = new Acquaintance(name, iban, publicKey, multiChain);
+
     }
 
-    /**
-     * getName() getter method testing
-     * @throws Exception
-     */
     @Test
-    public void getNameTest() throws Exception {
-        assertEquals(name, user.getName());
-    }
-
-    /**
-     * getIban() getter method testing
-     * @throws Exception
-     */
-    @Test
-    public void getIbanTest() throws Exception {
-        assertEquals(iban, user.getIban());
-    }
-
-    /**
-     * getPublicKey() getter method testing
-     * @throws Exception
-     */
-    @Test
-    public void getPublicKeyTest() throws Exception {
-        assertEquals(publicKey, user.getPublicKey());
+    public void getMultichain() throws Exception {
+        assertEquals(multiChain, acquaintance.getMultichain());
     }
 
 }
