+++ app/src/test/java/nl/tudelft/b_b_w/blockchain/BlockTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.b_b_w.blockchaincomponents;
+package nl.tudelft.b_b_w.blockchain;
 
 import net.i2p.crypto.eddsa.EdDSAPrivateKey;
 import net.i2p.crypto.eddsa.EdDSAPublicKey;
