+++ app/src/test/java/nl/tudelft/b_b_w/blockchain/HashTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.b_b_w.blockchaincomponents;
+package nl.tudelft.b_b_w.blockchain;
 
 import org.junit.Before;
 import org.junit.Test;
