+++ app/src/test/java/nl/tudelft/b_b_w/blockchain/UserTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.b_b_w.blockchaincomponents;
+package nl.tudelft.b_b_w.blockchain;
 
 import net.i2p.crypto.eddsa.EdDSAPrivateKey;
 import net.i2p.crypto.eddsa.EdDSAPublicKey;
@@ -7,16 +7,14 @@ import net.i2p.crypto.eddsa.Utils;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.util.ArrayList;
-
 import nl.tudelft.b_b_w.controller.ED25519;
 
 import static junit.framework.Assert.assertEquals;
 
 /**
- * Testing class for Acquaintance.java
+ * This class tests the User object.
  */
-public class AcquaintanceTest {
+public class UserTest {
 
     /**
      * The user attributes
@@ -24,9 +22,7 @@ public class AcquaintanceTest {
     private String name;
     private String iban;
     private EdDSAPublicKey publicKey;
-    private ArrayList<Chain> multiChain;
-
-    private Acquaintance acquaintance;
+    private User user;
 
     /**
      * Setting up the test object
@@ -37,15 +33,34 @@ public class AcquaintanceTest {
         iban = "NL642335674446";
         EdDSAPrivateKey edDSAPrivateKey = ED25519.generatePrivateKey(Utils.hexToBytes("0000000000000000000000000000000000000000000000000000000000000000"));
         publicKey = ED25519.getPublicKey(edDSAPrivateKey);
-
-        multiChain = new ArrayList<Chain>();
-        acquaintance = new Acquaintance(name, iban, publicKey, multiChain);
-
+        user = new User(name, iban, publicKey);
     }
 
+    /**
+     * getName() getter method testing
+     * @throws Exception
+     */
     @Test
-    public void getMultichain() throws Exception {
-        assertEquals(multiChain, acquaintance.getMultichain());
+    public void getNameTest() throws Exception {
+        assertEquals(name, user.getName());
+    }
+
+    /**
+     * getIban() getter method testing
+     * @throws Exception
+     */
+    @Test
+    public void getIbanTest() throws Exception {
+        assertEquals(iban, user.getIban());
+    }
+
+    /**
+     * getPublicKey() getter method testing
+     * @throws Exception
+     */
+    @Test
+    public void getPublicKeyTest() throws Exception {
+        assertEquals(publicKey, user.getPublicKey());
     }
 
 }
